Are-you-sure menu for root processes in killmenu.pl

+++ perl/killmenu.pl
@@ -16,11 +16,18 @@ $Data::Dumper::Sortkeys = 1;
 my $data = `ps aux`; # Already returns newline-delimited output
 my $selection = MenuSuite::promptMenu("Process List: ", $data) || exit 0;
 
-my $pid = (split /\W+/, $selection)[1];
+my @tokens = split /\W+/, $selection;
+
+my $user = $tokens[0];
+my $pid = $tokens[1];
 
 if (looks_like_number($pid))
 {
+    if ($user eq 'root')
+    {
+        my $check = MenuSuite::promptMenu("Killing a root process, Are you sure? (y/n) ", $selection) || exit 0;
+        exit 0 if uc($check) ne 'Y';
+    }
+
     system("kill $pid");
 }
-
-
